Add ContextManager error module
Moves the errors used in the ContextManager to a separate module.
